Public Cache interfaces

+++ apps/files_sharing/tests/backend.php
@@ -58,8 +58,10 @@ class Test_Files_Sharing_Backend extends TestCase {
 	}
 
 	protected function tearDown() {
-		$this->view->unlink($this->filename);
-		$this->view->deleteAll($this->folder);
+		if ($this->view) {
+			$this->view->unlink($this->filename);
+			$this->view->deleteAll($this->folder);
+		}
 
 		parent::tearDown();
 	}
